First steps to a full structure and code refactoring.
