+++ .gitignore
@@ -396,3 +396,7 @@ FodyWeavers.xsd
 
 # JetBrains Rider
 *.sln.iml
+
+# JetBRains Ide
+.idea/
+.junie/
