+++ manifest.json
@@ -6,6 +6,6 @@
     "js": "index.js",
     "css": "style.css",
     "author": "Rendal",
-    "version": "1.8",
+    "version": "2",
     "homePage": "https://github.com/sakhavhyand/SillyTavern-AnotherCharManager"
 }
