+++ modal.html
@@ -129,9 +129,9 @@
                     <hr>
                 </div>
             </div>
-            <div id="acm_popup_controls">
-                <div id="acm_switch_ui" class="acm_popup_buttons menu_button">Switch UI</div>
-                <div id="acm_popup_close" class="menu_button">Close</div>
+            <div class="popup-controls">
+                <div id="acm_switch_ui" class="menu_button">Switch UI</div>
+                <div id="acm_popup_close" class="menu_button popup-button-ok">Close</div>
             </div>
         </div>
     </div>
