+++ src/css/modal.css
@@ -1,0 +1,119 @@
+/* ===== MODAL COMPONENTS ===== */
+/* Main popup and overlay */
+#acm_shadow_popup {
+    backdrop-filter: blur(calc(var(--SmartThemeBlurStrength) * 2));
+    -webkit-backdrop-filter: blur(calc(var(--SmartThemeBlurStrength) * 2));
+    background-color: var(--black30a);
+    display: none;
+    opacity: 0.0;
+    position: absolute;
+    width: 100%;
+    height: 100svh;
+    z-index: 9000;
+    top: 0;
+}
+
+#acm_popup {
+    width: 500px;
+    max-width: 90svw;
+    position: absolute;
+    z-index: 9999;
+    margin-left: auto;
+    margin-right: auto;
+    left: 0;
+    right: 0;
+    top: 50%;
+    transform: translateY(-50%);
+    text-align: center;
+    box-shadow: 0 0 14px var(--black70a);
+    border: 1px solid var(--SmartThemeBorderColor);
+    padding: 4px;
+    background-color: var(--SmartThemeBlurTintColor);
+    border-radius: 10px;
+    max-height: 90svh;
+    display: flex;
+    flex-direction: column;
+    overflow-y: hidden;
+}
+
+#acm_popup_holder {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    overflow-y: hidden;
+    padding: 0 10px;
+}
+
+#acm_popup_text {
+    flex-grow: 1;
+    overflow-y: auto;
+    height: 100%;
+}
+
+/* Character popup */
+#acm_character_popup {
+    display: none;
+    background-color: var(--SmartThemeBlurTintColor);
+    backdrop-filter: blur(calc(var(--SmartThemeBlurStrength)*2));
+    -webkit-backdrop-filter: blur(calc(var(--SmartThemeBlurStrength)*2));
+    min-width: 100px;
+    max-width: var(--sheldWidth);
+    height: calc(100dvh - 84px);
+    min-height: calc(100dvh - 84px);
+    max-height: calc(100dvh - 84px);
+    position: absolute;
+    z-index: 9001;
+    margin-left: auto;
+    margin-right: auto;
+    left: 0;
+    right: 0;
+    top: var(--topBarBlockSize);
+    box-shadow: 0 0 20px var(--black70a);
+    padding: 10px;
+    border: 1px solid var(--SmartThemeBorderColor);
+    border-radius: 0 0 10px 10px;
+    overflow-y: auto;
+}
+
+#acm_character_popup h5 {
+    color: var(--grey50);
+}
+
+#acm_character_popup h3 {
+    margin-top: 0;
+    margin-bottom: 0;
+    margin-left: 0;
+    display: inline;
+}
+
+#acm_character_popup_text {
+    align-items: center;
+}
+
+#acm_character_cross {
+    position: absolute;
+    right: 5px;
+    top: 5px;
+    cursor: pointer;
+    opacity: 0.6;
+    font-size: 24px;
+}
+
+#acm_character_popup_ok {
+    cursor: pointer;
+    display: none;
+    height: 40px;
+    margin-top: 15px;
+    margin-left: 36px;
+    backdrop-filter: blur(var(--SmartThemeBlurStrength));
+    background-color: var(--black70a);
+    -webkit-backdrop-filter: blur(var(--SmartThemeBlurStrength));
+    width: 110px;
+    text-align: center;
+}
+
+/* Export format popup */
+#acm_export_format_popup {
+    display: none;
+    z-index: 9100;
+}
