+++ src/css/presets.css
@@ -1,0 +1,28 @@
+/* ===== CATEGORY COMPONENTS ===== */
+.acm_catCreate {
+    display: flex;
+    align-items: center;
+}
+
+.acm_catList {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: baseline;
+    background-image: linear-gradient(348deg, var(--white30a)2%, var(--grey30a)10%, var(--black70a)95%, var(--SmartThemeQuoteColor)100%);
+    margin-bottom: 5px;
+    border-radius: 10px;
+    padding: 0 5px;
+    border: 1px solid var(--SmartThemeBorderColor);
+}
+
+.acm_catTagList {
+    display: flex;
+    margin-bottom: 5px;
+    margin-left: 5px;
+    margin-right: 5px;
+}
+
+#preset_selector {
+    padding-right: 20px;
+}
