+++ src/css/tags.css
@@ -1,0 +1,43 @@
+/* ===== TAG COMPONENTS ===== */
+#tags-list {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.2rem;
+    height: 0;
+    overflow-y: auto;
+    margin-bottom: 5px;
+    transition: height 0.3s ease-out;
+}
+
+.acm_tag {
+    border-style: solid;
+    border-width: 1px;
+    border-radius: 5px;
+    box-sizing: border-box;
+    color: var(--SmartThemeBodyColor);
+    background-color: var(--black30a);
+    border-color: var(--white50a);
+    padding: 0.1rem 0.2rem;
+    font-size: calc(var(--mainFontSize) - 5%);
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    position: relative;
+    gap: 10px;
+    width: fit-content;
+    min-width: 0;
+    cursor: pointer;
+    opacity: 0.6;
+    filter: brightness(0.8);
+}
+
+.acm_tag_name {
+    text-overflow: ellipsis;
+    overflow: hidden;
+    text-align: left;
+    white-space: nowrap;
+}
+
+.tag-searchbar {
+    width: 50%;
+}
